Implemented template based implementation code generation

+++ cssfinder/algorithm/backend/numpy/_impl/_complex64.py
@@ -23,9 +23,9 @@
 Spec
 ----
 
-- Primary precision:    np.complex128
-- Floating precision:   np.float64
-- Complex precision:    np.complex128
+- Primary precision:    np.complex64
+- Floating precision:   np.float32
+- Complex precision:    np.complex64
 
 """
 
@@ -38,12 +38,12 @@ from numba import jit
 
 @jit(forceobj=True)
 def optimize_d_fs(
-    rho2: npt.NDArray[np.complex128],
-    rho3: npt.NDArray[np.complex128],
+    rho2: npt.NDArray[np.complex64],
+    rho3: npt.NDArray[np.complex64],
     depth: int,
     quantity: int,
     epochs: int,
-) -> npt.NDArray[np.complex128]:
+) -> npt.NDArray[np.complex64]:
     """Optimize implementation for FSnQd mode."""
 
     product_2_3 = product(rho2, rho3)
@@ -68,13 +68,13 @@ def optimize_d_fs(
             rotated_2 = rotate(rotated_2, unitary)
             product_rot2_3 = product(rotated_2, rho3)
 
-    return rotated_2.astype(np.complex128, copy=False)  # type: ignore
+    return rotated_2.astype(np.complex64, copy=False)  # type: ignore
 
 
 @jit(nopython=True, nogil=True, cache=True)
 def product(
-    matrix1: npt.NDArray[np.complex128], matrix2: npt.NDArray[np.complex128]
-) -> np.float64:
+    matrix1: npt.NDArray[np.complex64], matrix2: npt.NDArray[np.complex64]
+) -> np.float32:
     """Calculate scalar product of two matrices."""
 
     return np.trace(np.dot(matrix1, matrix2)).real  # type: ignore
@@ -83,7 +83,7 @@ def product(
 @jit(forceobj=True, cache=True)
 def random_unitary_d_fs(
     depth: int, quantity: int, idx: int
-) -> npt.NDArray[np.complex128]:
+) -> npt.NDArray[np.complex64]:
     """N quDits."""
     value = _random_unitary_d_fs_val(depth)
     mtx = expand_d_fs(value, depth, quantity, idx)
@@ -97,7 +97,7 @@ _VALUE = _REAL + _IMAG - 1
 
 
 @jit(nopython=True, nogil=True, cache=True)
-def _random_unitary_d_fs_val(depth: int) -> npt.NDArray[np.complex128]:
+def _random_unitary_d_fs_val(depth: int) -> npt.NDArray[np.complex64]:
     random_mtx = random_d_fs(depth, 1)
     identity_mtx = np.identity(depth)
 
@@ -109,7 +109,7 @@ def _random_unitary_d_fs_val(depth: int) -> npt.NDArray[np.complex128]:
 
 
 @jit(nopython=True, nogil=True, cache=True)
-def random_d_fs(depth: int, quantity: int) -> npt.NDArray[np.complex128]:
+def random_d_fs(depth: int, quantity: int) -> npt.NDArray[np.complex64]:
     """Random n quDit state."""
     vector = normalize(get_random_haar(depth))
 
@@ -123,17 +123,17 @@ def random_d_fs(depth: int, quantity: int) -> npt.NDArray[np.complex128]:
 
 
 @jit(nopython=True, nogil=True, cache=True)
-def get_random_haar(depth: int) -> npt.NDArray[np.complex128]:
+def get_random_haar(depth: int) -> npt.NDArray[np.complex64]:
     """Generate a random vector with Haar measure."""
     real = np.random.normal(0, 1, depth)
     imaginary = np.random.normal(0, 1, depth)
     # Complex128 is a correct type returned from this expression.
     # Checked on numpy 1.23.5
-    return (real + 1j * imaginary).astype(np.complex128)  # type: ignore
+    return (real + 1j * imaginary).astype(np.complex64)  # type: ignore
 
 
 @jit(nopython=True, nogil=True, cache=True)
-def normalize(mtx: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
+def normalize(mtx: npt.NDArray[np.complex64]) -> npt.NDArray[np.complex64]:
     """Normalization of a vector."""
 
     mtx2 = np.dot(mtx, np.conj(mtx))
@@ -144,7 +144,7 @@ def normalize(mtx: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
 
 
 @jit(nopython=True, nogil=True, cache=True)
-def project(mtx1: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
+def project(mtx1: npt.NDArray[np.complex64]) -> npt.NDArray[np.complex64]:
     """Build a projection from a vector."""
 
     retval = np.outer(mtx1, np.conj(mtx1))
@@ -154,31 +154,31 @@ def project(mtx1: npt.NDArray[np.complex128]) -> npt.NDArray[np.complex128]:
 
 @jit(forceobj=True, cache=True)
 def expand_d_fs(
-    value: npt.NDArray[np.complex128],
+    value: npt.NDArray[np.complex64],
     depth: int,
     quantity: int,
     idx: int,
-) -> npt.NDArray[np.complex128]:
+) -> npt.NDArray[np.complex64]:
     """Expand an operator to n quDits."""
 
     depth_1 = int(depth**idx)
-    identity_1 = np.identity(depth_1).astype(np.complex128, copy=False)
+    identity_1 = np.identity(depth_1).astype(np.complex64, copy=False)
 
     depth_2 = int(depth ** (quantity - idx - 1))
-    identity_2 = np.identity(depth_2).astype(np.complex128, copy=False)
+    identity_2 = np.identity(depth_2).astype(np.complex64, copy=False)
 
     kronecker_1 = kronecker(identity_1, value)
     kronecker_2 = kronecker(kronecker_1, identity_2)
 
-    retval = kronecker_2.astype(np.complex128, copy=False)
+    retval = kronecker_2.astype(np.complex64, copy=False)
 
     return retval  # type: ignore
 
 
 @jit(forceobj=True, cache=True)
 def kronecker(
-    mtx: npt.NDArray[np.complex128], mtx1: npt.NDArray[np.complex128]
-) -> npt.NDArray[np.complex128]:
+    mtx: npt.NDArray[np.complex64], mtx1: npt.NDArray[np.complex64]
+) -> npt.NDArray[np.complex64]:
     """Kronecker Product."""
     ddd1 = len(mtx)
     ddd2 = len(mtx1)
@@ -187,15 +187,15 @@ def kronecker(
     dot_0_1 = np.tensordot(mtx, mtx1, 0)
     out_mtx = np.swapaxes(dot_0_1, 1, 2)
 
-    retval = out_mtx.reshape(output_shape).astype(np.complex128, copy=False)
+    retval = out_mtx.reshape(output_shape).astype(np.complex64, copy=False)
 
     return retval  # type: ignore
 
 
 @jit(nopython=True, nogil=True, cache=True)
 def rotate(
-    rho2: npt.NDArray[np.complex128], unitary: npt.NDArray[np.complex128]
-) -> npt.NDArray[np.complex128]:
+    rho2: npt.NDArray[np.complex64], unitary: npt.NDArray[np.complex64]
+) -> npt.NDArray[np.complex64]:
     """Sandwich an operator with a unitary."""
 
     rho2a = np.dot(rho2, np.conj(unitary).T)  # matmul replaced with dot
